Synchronise the customname with the shelly name
The /CustomName is used in the devicepage and /SwitchableOutput/0/Settings/CustomName is used in the switch pane.
Since each channel gets its own service, it makes sense to use the same name for both.
When users give a shelly channel a descriptive name for the application, it will show up on the device list and on the switch pane.